Accumulation is set per table, not per update op.

+++ src/sparrow/util/file.cc
@@ -17,7 +17,7 @@ vector<string> File::MatchingFilenames(StringPiece pattern) {
   globbuf.gl_offs = 0;
   glob(pattern.AsString().c_str(), 0, NULL, &globbuf);
   vector<string> out;
-  for (int i = 0; i < globbuf.gl_pathc; ++i) {
+  for (size_t i = 0; i < globbuf.gl_pathc; ++i) {
     out.push_back(globbuf.gl_pathv[i]);
   }
   globfree(&globbuf);
@@ -27,7 +27,7 @@ vector<string> File::MatchingFilenames(StringPiece pattern) {
 vector<File::Info> File::MatchingFileinfo(StringPiece glob) {
   vector<string> names = MatchingFilenames(glob);
   vector<File::Info> out(names.size());
-  for (int i = 0; i < names.size(); ++i) {
+  for (size_t i = 0; i < names.size(); ++i) {
     out[i].name = names[i];
     stat(names[i].c_str(), &out[i].stat);
   }
@@ -44,7 +44,7 @@ void File::Mkdirs(string path) {
 
   vector<StringPiece> pbits = StringPiece::split(path, "/");
   string prefix;
-  for (int i = 0; i < pbits.size(); ++i) {
+  for (size_t i = 0; i < pbits.size(); ++i) {
     pbits[i].strip();
     if (pbits[i].size() == 0) {
       continue;
